Reorganized src files, cleaned up manual tests
