+++ manual_tests/can_commands_test/can_commands_test.c
@@ -1,0 +1,7 @@
+/*
+An outdated test for simulating EPS receiving and transmitting CAN messages.
+TODO - update this test, similar to PAY
+*/
+
+int main(void) {
+}
