+++ manual_tests/can_commands_test/makefile
@@ -1,4 +1,4 @@
-PROG = can_test
+PROG = can_commands_test
 # SRC should only include necessary files
 #SRC = $(addprefix ../../src/,sensors.c pex.c)
 include ../makefile
