+++ manual_tests/shunts_test/makefile
@@ -1,3 +1,3 @@
 PROG = shunts_test
-SRC = $(addprefix ../../src/,shunts.c)
+SRC = $(addprefix ../../src/,devices.c shunts.c)
 include ../makefile
