+++ src/can_interface.h
@@ -1,5 +1,5 @@
-#ifndef CAN_CALLBACKS_H
-#define CAN_CALLBACKS_H
+#ifndef CAN_INTERFACE_H
+#define CAN_INTERFACE_H
 
 #include <stdint.h>
 #include <can/can.h>
@@ -8,8 +8,7 @@
 #include <queue/queue.h>
 #include <uart/uart.h>
 
-extern queue_t can_rx_msgs;
-extern queue_t can_tx_msgs;
+#include "can_commands.h"
 
 // CAN MOBs
 extern mob_t cmd_rx_mob;
