+++ src/devices.h
@@ -1,0 +1,12 @@
+#ifndef DEVICES_H
+#define DEVICES_H
+
+#include <adc/adc.h>
+#include <adc/eps.h>
+#include <pex/eps.h>
+#include <pex/pex.h>
+
+extern adc_t adc;
+extern pex_t pex;
+
+#endif
