+++ src/shunts.h
@@ -2,13 +2,15 @@
 #define SHUNTS_H
 
 #include <stdbool.h>
-#include <pex/pex.h>
-#include <pex/eps.h>
+
 #include <adc/adc.h>
 #include <adc/eps.h>
 #include <conversions/conversions.h>
+#include <pex/eps.h>
+#include <pex/pex.h>
 #include <uart/uart.h>
-#include "eps.h"
+
+#include "general.h"
 
 // If the battery output voltage goes above this threshold,
 // the shunts should be turned on (battery charging off)
